+++ Source/Shaders/BillboardCollectionVS.glsl
@@ -46,20 +46,32 @@ void main()
     ///////////////////////////////////////////////////////////////////////////     
     
 #ifdef EYE_DISTANCE_SCALING  // scale based on eye distance
-    // transform into 3d eye coordinates.  This methodology will even translate for
-    // 2D mode into a 3D position, but is only valid for the distance calculations below
-    vec4 position3dInEC = czm_view3D * (vec4(positionHigh, 1) + vec4(positionLow,0));
-    float lengthSq = dot(position3dInEC.xyz, position3dInEC.xyz);
+    float lengthSq;
+    if (czm_sceneMode == czm_sceneMode2D)
+    {
+        // 2D camera distance is a special case
+        // treat all billboards as flattened to the z=0.0 plane
+        lengthSq = czm_eyeAltitude2D * czm_eyeAltitude2D;
+    }
+    else
+    {
+        lengthSq = dot(positionEC.xyz, positionEC.xyz);
+    }
+
     float scaleAtMin = scaleByDistance.y;
     float scaleAtMax = scaleByDistance.w;
-    float minDistanceFromEyeSq = scaleByDistance.x*scaleByDistance.x;
-    float maxDistanceFromEyeSq = scaleByDistance.z*scaleByDistance.z;
+    float nearDistanceSq = scaleByDistance.x*scaleByDistance.x;
+    float farDistanceSq = scaleByDistance.z*scaleByDistance.z;
     // ensure that t will fall within the range of [0.0, 1.0]
-    lengthSq = clamp(lengthSq, minDistanceFromEyeSq, maxDistanceFromEyeSq);
-    float t = (lengthSq-minDistanceFromEyeSq)/(maxDistanceFromEyeSq-minDistanceFromEyeSq);
+    lengthSq = clamp(lengthSq, nearDistanceSq, farDistanceSq);
+
+    float t = (lengthSq-nearDistanceSq)/(farDistanceSq-nearDistanceSq);
+    // our exponent used to dampen the interpolation is tied to distance, where greater distances
+    // will yield smaller exponents, and thus more gradual interpolation
+    float x = (farDistanceSq-lengthSq)/farDistanceSq;
     // dampen the exponential distance traveled by the camera
     // could be interesting to assign this exponent to a uniform for advanced control of interpolation
-    t = pow(t, 0.5);
+    t = pow(t, x);
 
     scale *= mix(scaleAtMin, scaleAtMax, t);
 #endif